# cmd/apps/apis/rest/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  cmd_apps_apis_rest_v1.DocumentProperty:
    properties:
      abstract:
        type: string
      author:
        type: string
      header_image:
        type: string
      keywords:
        items:
          type: string
        type: array
      marked:
        type: boolean
      notes:
        type: string
      publish_at:
        type: string
      source:
        type: string
      title:
        type: string
      unread:
        type: boolean
      url:
        type: string
      year:
        type: string
    type: object
...
  cmd_apps_apis_rest_v1.EntryInfo:
    properties:
      access_at:
        type: string
      changed_at:
        type: string
      created_at:
        type: string
      document:
        $ref: '#/definitions/cmd_apps_apis_rest_v1.DocumentProperty'
      entry:
        type: integer
      is_group:
        type: boolean
      kind:
        type: string
      modified_at:
        type: string
      name:
        type: string
      size:
        type: integer
      uri:
        type: string
    type: object
...
  cmd_apps_apis_rest_v1.ListEntriesResponse:
    properties:
      entries:
        items:
          $ref: '#/definitions/cmd_apps_apis_rest_v1.EntryInfo'
        type: array
      pagination:
        $ref: '#/definitions/v1.PaginationInfo'
    type: object
...
  cmd_apps_apis_rest_v1.WriteFileResponse:
    properties:
      len:
        type: integer
    type: object
...
  v1.GroupEntry:
    properties:
      children:
        items:
          $ref: '#/definitions/v1.GroupEntry'
        type: array
      name:
        type: string
      uri:
        type: string
    type: object
  v1.GroupTreeResponse:
    properties:
      root:
        $ref: '#/definitions/v1.GroupEntry'
    type: object
...
  v1.PaginationInfo:
    properties:
      page:
        type: integer
      page_size:
        type: integer
    type: object
...
host: localhost:7081
info:
  contact: {}
  description: NanaFS is a Reference Filing System API
  title: NanaFS API
  version: "1.0"
paths:
...
  /api/v1/files/content:
    get:
      consumes:
      - application/json
      description: Read file content
      parameters:
      - description: File URI
        in: query
        name: uri
        type: string
      - description: File ID
        in: query
        name: id
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: File content
          schema:
            type: string
      summary: Read file
      tags:
      - Files
    post:
      consumes:
      - multipart/form-data
      description: Write file content via multipart upload
      parameters:
      - description: File URI
        in: query
        name: uri
        type: string
      - description: File ID
        in: query
        name: id
        type: string
      - description: File content
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/cmd_apps_apis_rest_v1.WriteFileResponse'
      summary: Write file
      tags:
      - Files
  /api/v1/groups/children:
    get:
      consumes:
      - application/json
      description: List children entries of a group with pagination
      parameters:
      - description: Group URI
        in: query
        name: uri
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: page_size
        type: integer
      - description: Order field
        in: query
        name: order
        type: string
      - description: Sort descending
        in: query
        name: desc
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/cmd_apps_apis_rest_v1.ListEntriesResponse'
      summary: List group children
      tags:
      - Groups
  /api/v1/groups/tree:
    get:
      consumes:
      - application/json
      description: Retrieve the complete group tree structure
      parameters:
      - description: Group URI
        in: query
        name: uri
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.GroupTreeResponse'
      summary: Get group tree
      tags:
      - Groups
...
swagger: "2.0"